Add knowledge base of the PALOMERA project (OA books policies)
PALOMERA: Policy Alignment of Open access Monographs in the European
Research Area

«  The Knowledge Base is a collection of documents, such as reports,
policies, survey results and statistics, relevant to Open Access (OA)
policies regarding OA books in the European Research Area. The
collection was created as part of the PALOMERA project. In a nutshell,
PALOMERA seeks to understand why so few OA funder policies include
books, and to provide actionable recommendations to change this.»